realtime-streaming-pipeline/requirements.txt
confluent-kafka==2.3.0
pyspark==3.5.0
delta-spark==3.0.0
boto3==1.34.0
redshift-connector==2.1.3
aws-cdk-lib==2.130.0
constructs==10.3.0
apache-airflow==2.8.0
apache-airflow-providers-amazon==8.18.0
python-dotenv==1.0.0
